requirements.txt
pyautogen>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0